+++ src/Domain/Models/Results/ChallengeSearchResultItem.cs
@@ -7,6 +7,6 @@ public class ChallengeSearchResultItem
     public string Id { get; set; }
     public string Name { get; set; }
     public string Task { get; set; }
-    public int Difficulty { get; set; }
+    public double Difficulty { get; set; }
     public IEnumerable<TagEntity> Tags { get; set; }
 }
